Remove unused parameters of global config update
Remove the parameters because:
 starting is always null.
 deferResume is always false.

Some update methods return whether the top activity is relaunched by
the config change. But there is no one check that. All callers check
whether the config is changed. So the legacy UpdateConfigurationResult
can be removed. This avoids confusing about different meaning of the
returned boolean from the methods with the same name.

Bug: 258618073
Bug: 364879728
Flag: EXEMPT remove dead code
Test: ActivityTaskManagerServiceTests# \
      testPackageConfigUpdate_locales_successfullyApplied

Change-Id: I68ec1ea49815234c6225ed37192645fa0a441e3c