[PiP2] Collect PiP candidate parent before request
We want to collect the parent of the PiP candidate activity
upon enterPictureInPictureMode() before sending requestStartTransition
for TRANSIT_PIP to Shell, since it's participating in TRANSIT_PIP in any
case.

 - When do issues arise that we are fixing here?

If the previous transition is, for example, transient-hiding
the parent of the pip candidate and actually HIDING the task upon its finishTx,
then checkEnterPipOnFinish passes in finishTransition() and doesn't
commitVisibility for that task.

This means WindowAnimator can come in and show() that task's leash
before TRANSIT_PIP formally starts.

This can be avoided if WindowAnimator recognizes the task as
participating in a collecting transition, letting it handle the
visibility.

Bug: 421969409
Flag: com.android.wm.shell.enable_pip2
Test: manually reproduce the steps in the bug
Change-Id: Iffba998b9cd8e47ba4b0f098ed8ef4d102309fcc