Persist desktops on external displays.
Stores uniqueDisplayId in DesktopPersistentRepository so when the
repository is initialized, the restored desks can be added to
repositories on displays other than the default.

Adds a method to DisplayController to search for a Display by uniqueId
to accomplish this.

Additionally updates instances of wct.launchTask to add appropriate
displayId to the ActivityOptions to ensure background tasks are not
launched on default display.

Bug: 427777565
Test: Manual; make desks on multiple displays and reboot; confirm desks
appear in correct display in overview.
Flag: com.android.window.flags.enable_display_reconnect_interaction
Change-Id: I638453b4ba0bf42781676c1e3d172d0dbbbe327d

Change-Id: Ib0d9cc090e8b3dbda1bf5c8164775c3d8815ad3f