Always dispatch tool type for showing IME
When we are already showing the IME and we are getting a showSoftInput
request, it's still possible that we need to dispatch the tool type, as
the tool type can change. This patch makes sure we call report the tool
type even if we don't want to call showSoftInput.

Fix: 425857347
Test: atest android.view.inputmethod.cts.StylusHandwritingTest#testOnViewClicked_withFingerTap
Flag: android.view.inputmethod.report_animating_insets_types
Change-Id: I4c9f69accb56cfe05671a5f838739473df727de9