HSU is not an Admin
We no longer mark the Headless System User as being an Admin.

The concept of Admin is supposed to be at the UI level - whether we
present the user with certain abilities - and the HSU, as the least
trustworthy user, should not have those abilities. The concept of Admin
should't make  difference at the actual service level, where checks
should involve UserRestrictions and permissions, not the UI-relevant
Admin flag.

Test: manual
Bug: 419105275
Flag: android.multiuser.hsu_not_admin
Change-Id: Ib6fc535b0b60547605ee66c2f3102f26ef31c425