Lockscreen wallpaper is blurred after switching back from Hub
The window remains blurred after transitioning from Hub to AOD and then
returning to lockscreen. Fix is to add the exit blur flow to the ViewModel.

Also fixes that after AOD->Hub, the hub's background isn't blurred when
it should be. It's because we forgot to bind the view model in the hub
transition module.

Fixes: 419426901
Test: atest GlanceableHubToAodTransitionViewModelTest
Test: atest AodToGlanceableHubTransitionViewModelTest
Flag: com.android.systemui.glanceable_hub_blurred_background
Change-Id: I3c5dc99dc4697121bfd8c1da6cd5b5dfce71d699