Protect virtual cameras on default policy with a flag
Add `external_camera_default_policy` flag to separate adding cameras on
the default device from the rest of the virtual external cameras support
and virtual camera metadata feature.

This is to decouple and separate any privacy concerns for the part that
still needs to be discussed by pwg when they have time from the rest of
the feature for which there are no concerns.

Bug: 431724713
Bug: 371167033
Test: atest VirtualCameraControllerTest
Test: atest VirtualCameraTest
Flag: android.companion.virtualdevice.flags.external_camera_default_policy
Change-Id: I8b1142d7110c9c7bdcc99c2e536a7682981452b3