Inactive NSSL for touch when blurred out on lockscreen
Disable NSSL from handling touch events when it's blurred out on
lockscreen. Instead, we want the flexi scene container to handle
the touches. Note that we need to make further changes if we
decide to show HUNs on this state as currently we disabled all
notifications from being clickable at this state.

When interactive is false, NSSL and descendants will not be
focusable or selectable by a11y.

Fix: 420601264
Test: manual
Flag: com.android.systemui.scene_container
Change-Id: Id5d6eebd4490cd4c68a91e0981bf427aa4e7093e