Fix Flaky Test
The goal of the vibrate_singleVibratorRepeatingAlwaysOnWaveform_turnsVibratorBackOn test is to verify that the second segment runs for it full intended duration, even after being cancelled mid step. Therefore, we've updated the condition to greater than or equal. This change handles a race condition where the device causes the segment to run for exactly the expected duration, rather than slightly longer.

Fix: 431925843
Flag: TEST_ONLY
Test: atest VibrationThreadTest#vibrate_singleVibratorRepeatingAlwaysOnWaveform_turnsVibratorBackOn
Change-Id: I55906239fc58667e0b3e895adfc71f838624f323