Remove synchronous startInputOrWindowGainedFocus
This removes the synchronous variant of startInputOrWindowGainedFocus,
resolving to the asynchronous method instead. The async one was
originally added in [1] as part of zero jank proxy. The flag was later
removed in [2], but the synchronous variant was kept due to one last
usage with WINDOW_FOCUS_GAIN_REPORT_ONLY.

This last synchronous call is now changed to be async, with the two
variants merged. Additionally, the aidl method is now marked oneway.

  [1]: I37fd585eeccab7c3b18fb509b3607c56d800c306
  [2]: Idbbb3c64511cb38275a4ad9c2bd7e33299ef4b56

Flag: EXEMPT cleanup
Test: atest CtsInputMethodTestCases
Bug: 418839448
Change-Id: Idf95bdd274a78243a86958286d9ba6a140957515