Add /vendor/odm/overlay to framework RRO allowlist
Framework RROs were allowed in vendor or odm partition.  However, they
were not allowed the odm subdirectory of the vendor partition which is
used by devices without a separate odm partition.  This difference was
not intended so add the /vendor/odm/overlay path to the allowed list.

Change-Id: I55b618320ed5b9f599173d4377189d0b67b1e836
Flag: EXEMPT Bugfix
Bug: 432463913
Test: install odm RRO on device without odm partition, see it enabled