Create the flag and the stubs for the App Lock APIs
Create the feature flag for the App Lock APIs. When this flag is set to
true, access to the APIs will be enabled. All callers of the APIs must
first check that the value of this flag is true, and when this flag is
false, the APIs will be excluded from public SDK and documentation. Also
create the API stubs PackageManager and the LOCK_APPS permission.
Creating the API feature flag requires creating the flagged APIs. Add
an OWNERS file for future App Lock changes.

Flag: android.security.app_lock_apis
Bug: 415167460
API-Coverage-Bug: 435250591
Test: Flag definition
Change-Id: Iccfd2b8bbe24d2ffe22040fc46413fcb21d4f68e