Disable window and transition animations on computercontrol displays
Introduce a method in WindowManagerInternal to enable/disable window
and transiton animations for a display, and store a boolean in
DisplayContent to indicate on/off state of animations. VDM service
calls this method for every display created for computercontrol (this
will also be called when a ComputerControlSession is attached/detached
to a MirrorView to re-enable animations). WMCore notifies WMShell using
a new method in IDisplayWindowListener about animations turned on/off
for a display, to play transitions accordingly.

Test: atest WindowManagerServiceTests
Test: atest ComputerControlSessionTest
Bug: 427437362
Flag: android.companion.virtualdevice.flags.enable_animations_per_display
Change-Id: Ie4835c1107d33556c04ace1a59d7c7341b510e77