Ensure that a11y focus stays within Open By Default Dialog
The expected a11y behaviour for dialogs is for the a11y focus to remain
in the dialog. Since the OpenByDefaultDialog is a surface and not a
standalone window, this isn't handled by default, so to mimic the
experience, we can set the a11y traversal logic for each dialog element
to create a loop.

Bug: 398077805
Flag: EXEMPT bugfix
Test: manual (with talkback)
Change-Id: I1e46ed1c5c1cc979e78728d1d678b81674ef5d79