Remove locale and language from ImeSubtypeListItem
These properites were previously used in ordering the list of Input
Methods and Subtypes for the IME Switcher Menu, prioritizing those that
fit either the system language or system locale.

However in [1] we modified this to only sort based on the IME name, and
keep the order of subtypes per IME as given by the IMEs themselves.
Thus, these properites are now unused and can be removed.

  [1]: I2fba8c0ac6e756539deb433a77a8d647a30134ab

Flag: EXEMPT cleanup
Bug: 347693610
Test: atest InputMethodSubtypeSwitchingControllerTest
Change-Id: I508bc4c700df1c622a067c4e32d2b1a323fc882e