Add Lock global action to Power Menu
Add a Lock Power Menu action that does not require strong auth, as
opposed on the Lockdown action. This enables users to unlock with
fingerprint, for example. The new Lock action will only be applied
on desktop devices (via the desktop config.xml), where it will
replace the Lockdown button.

Bug: 393319004
Test: atest GlobalActionsDialogLiteTest
Test: manual verification on desktop device
Flag: NONE. The global action will only be configured for desktop
Change-Id: Idd124491c6b1ccf1368726b96b37974949e7d34b