[SB][ComposeIcons] Add a composable TTY icon
This is a reland of ag/34734530. The original CL had an
incorrect test in SystemStatusIconsViewModelTest

This CL adds a recommended architecture implementation
of the legacy implementation for the TTY icon in the
status bar. This includes a new interactor and viewmodel.

Bug: 433807489
Test: TtyIconViewModelTest, TtyStatusInteractorTest
Flag: com.android.systemui.status_bar_system_status_icons_in_compose
Change-Id: Ie274500b7a8b7d5f183a6a5dfc8d278076b37892