Clean up TelephonyRegistryManager Caching
For some reason there is a member variable of the
TelephonyRegistryManager that was being both cached
and retrieved every time it was being used. For
simplicity, for the time being at least, I am removing
the member variable since the code was already fetching
it. That makes this a non-functional change since it was
never actually checking whether the cached value was
already non-null (and why would it)?

Bug: 433336412
Test: atest TelephonyManagerTest
Flag: EXEMPT bugfix
Change-Id: I1ccf3de3b0dcfc7b01c264ef9283e3cd0c419f79