Remove redundant RuntimeExceptions from ProtoLogs
This is a follow up of [1], where some stack trace calls were replaced
with RuntimeExceptions. The intention was to simply use the exceptions
to directly get the stack trace, but this behaviour only works for
logcat, whereas ProtoLog ended up concatenating the string
"RuntimeException()" explicitly. This removes the exceptions as these
are no longer used.

  [1]: I46630d4ac860141f1142af509d24ee33568068d8

Flag: EXEMPT cleanup
Bug: 407383437
Test: presubmit
Change-Id: I1565f57b8d07e6b5df187c940890a9129f674f88