Store current carousel position
This CL stores the current index of media carousel to be shared across different locations as long as media carousel is visible to user. Once media is not showing, media carousel should reorder its media sessions and the current index should be 0 in order to make the carousel scroll to the first card in the current media list.

Flag: com.android.systemui.media_controls_in_compose
Bug: 397989775
Test: Checked UI.
Change-Id: I54e7f675631dfe7fa916a353a13de1d46b96cc29