Merge CLOSE transitions to unfold transitions
When we unfold while we hold the -1 screen on launcher
in slightly opened state, the TaskFragment will be dismissed
and hidden. This might happen when we already started animating
unfold shell transition, in this case the dismissal will
be animated only after the unfold shell transition is finished.

To fix this issue we could accept TRANSIT_CLOSE merge request
for the task fragment close transitions, and hide the participating
layers immediately as part of startTransaction.

Bug: 431983411
Flag: EXEMPT bugfix
Test: atest UnfoldTransitionHandlerTest
Change-Id: I0b870dfe080bd89ebbc6c00ec9de2704c2227ff6