BouncyBall: Rename package
We inherited the "com.prefabulated" from TouchLatency.  That's a
reference to an old engineering inside joke.

With the intention of this app being used widely, as part of
qualifying devices, we lose the humor to help assure clarity
and avoid confusion (especially with folks unfamiliar with the
joke).

There are already 44 test apps in this project using
"com.android.test", so we switch to that.

Test: Local build; try out README instructions
Bug: 408044970
Flag: EXEMPT for test app
Change-Id: Ieca6a4fe2ecc05180beec59487dde86ac967f7f7