Allowlist and advertise hidden routes when they are selected
If a media route is hidden from an app due to the app lacking required
permissions for the route, when the route is selected in a privileged
UI such as the output switcher, we need to allowlist the app to have
access to it. Also, the app may miss the transfer event we send if it
uses androidx wrappers because the androidx wrapper currently ignores
MediaRouter2.TransferCallback events for routes which haven't been
previously advertised via a MediaRouter2.RouteCallback event.

To prevent this problem, this CL adds code to advertise hidden routes
route just after allowlisting then, before sending any transfer
events.

Bug: 388071860
Test: atest MediaRouter2HostSideTest
Flag: com.android.media.flags.enable_route_visibility_control_compat_fixes
Change-Id: Iec6df8fc4b1959285fdd4187ee71388a7ddf6c0f