DreamOverlayService: Reduce scope where shade affects gesture blocking.
Currently shade expansion affects gesture blocking and is active during
the lifetime of the DreamOverlayService. Since a service can be reused
by multiple dream instances, it is better for this listener to be
scoped to the dream lifetime, rather than the overlay service.
Additionally, the callback for handling changes fires on an executor. As
a result, it might be executed after the dream has ended. This
changelists adds a scoped execution method which checks the dream
status before executing the Runnable.

Fixes: 432039567
Test: atest DreamOverlayServiceTest
Flag: EXEMPT bugfix
Change-Id: I82e20461a27956513aa5076bc003c636cfa30a2c