Show clipboard overlay on correct display
Sets the launch display id on the intents sent from clipboard to ensure
that they're launched on the right display (the same one as where the UI
appears/where the content was copied from) and gives the toast
controller (used if e.g. setup is incomplete) the context to show on the
correct display.

Also adds mLaunchDisplayId to the set of values that gets copied over in
ActivityOptions.update(otherOptions: ActivityOptions).

Bug: 425706893
Test: atest ClipboardOverlayController
Test: manual using virtual displays
Flag: com.android.systemui.clipboard_overlay_multiuser

Change-Id: I09027317550ec0bc2cb14e6e20d8ebdec2d0eb86