psc: Remove redundant uid field from ProcessRecord
The `uid` field in `ProcessRecord` was a duplication of the `mUid`
field already present in its superclass, `ProcessRecordInternal`. This
patch removes the redundant field from `ProcessRecord` and renames
`mUid` to `uid` in `ProcessRecordInternal`, making it public. This
consolidates the process UID definition in the base class.

Bug: 425766486
Test: m services.core
Test: atest MockingOomAdjusterTests OomAdjusterTests
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT pure refactor
Change-Id: Ie984ad388d0960486e2af6d03982c595408e4e50