Allow CMP exemption for restricted bucket due to reason timeout.
In case of media backup jobs, we want to run them even when the
app is in a restricted standby bucket if the reason is timed out
instead of some other reason. This is to ensure that backup jobs
are allowed to run even if the user doesn't manually use the app.

Bug: 409606405
Test: atest ./services/tests/mockingservicestests/src/com/android/server/job/controllers/*Test.java
Test: atest ./services/tests/mockingservicestests/src/com/android/server/job/*Test.java
Flag: com.android.server.job.allow_cmp_exemption_for_restricted_bucket
Change-Id: I12c8c7d941d57b2d0c57bf39596236506edbdb94