Fix foreground service visibility in Task Manager
Under certain conditions, a foreground service was not being correctly
identified by the system. This caused the service to be absent from the
Task Manager UI, preventing users from seeing or managing it.

This change corrects the service state management
to ensure that when a service is promoted to the foreground, its visible
in the TaskManager UI.

Flag: com.android.server.am.fgs_visibility_fix
Bug: 417784630
Test: atest ActivityManagerTest ActivityManagerServiceTest
Change-Id: I80de4dcf98d445737929c01dc534b172262c0e5e