[flexiglass] Replace ShadeRepository.isShadeLayoutWide with two flows.
This splits isShadeLayoutWide into "isWideScreen" (whose value is only
based on the current width of the screen that hosts the shade window),
and "legacyUseSplitShade" which is based on the config value
`config_use_split_notification_shade`. They differ by the fact that the
latter is "false" on large screen devices in portrait orientation.

This is part of a multi-CL refactor that is meant to disambiguate the
meaning of "shade layout wide" and simplify related decision-making
logic.

Bug: 354926927
Bug: 411365508
Test: Added unit test coverage.
Test: All existing unit tests pass.
Flag: com.android.systemui.scene_container
Change-Id: I9b8417e9e6784dab24833a4f845ba4d0950425d0