[Flexiglass] Fix crash in split shade
With the new media (and deleting unused code in flexiglass) I deleted
the initialization of the QQS media host. Even though this is not unsued
anywhere in flexilgass, QSImpl (what's still used for split shade),
still attaches all its views (including QQS) that expected an
initialized media host.

With this, we make sure to initialize it where it's needed, knowing that
this class will be deleted soon (after QSComposeFragment releases). This
also should not be an issue once QS in split shade is replaced with the
new version (and QSSceneAdapter and friends is deleted).

Test: manual, open split shade
Test: atest QuickQSPanelControllerTest
Fixes: 437073248
Flag: com.android.systemui.scene_container
Change-Id: Id7ba23ca2e02ce31f7c210372353eb9276b764a9