[flexiglass] Make the Bouncer bottom padding resize to available space.
The Emergency Action Button was being squished because of the high
bottom padding (48dp) required of it even when the IME is open. By
changing 16dp of that padding to collapse to zero when there isn't
sufficient vertical space, we're avoiding that problem.

This also reduces some of the padding between the emergency button and
the input field, such that the paddings around the button are more
visually balanced in that scenario.

Fix: 436813358
Test: Manually tested the password bouncer on an unfolded device.
Flag: com.android.systemui.scene_container
Change-Id: Ib6e0f43d76c612c49b865f6fdb890f62f0139503