fix(edt): Ensure the NavBar background is always dark when EDT is on
In non edge-to-edge apps which draw this background as a black rect,
expanded dark theme would incorrectly invert this background to appear
white. This then causes the NavBar buttons to appear invisible because
they are rendered by a separate process (Launcher) and are always light.

Bug: 435736433
Test: manual only; observe the NavBar stays dark, including for
      non-edge-to-edge apps, when rotating portrait and landscape,
      and in both 3-button and gesture nav
Flag: android.view.accessibility.force_invert_color
Change-Id: Id50d1394328fe40bbd37c9e75bddb7a3eb5cc0a0