Fix restoring custom ringtone with custom vibration
Currently, the restore process would be failed because the media file id
could be different in different devices, and that would cause the audio
file can't be found in the new device.

This CL will fix this by calling `getRingtoneUriForRestore' first to get
the restore uri for ringtone audio then append the restore uri for
vibration.

Bug: 436490960
Test: atest SettingsHelperTest
Flag: EXEMPT BUGFIX
Change-Id: I72029350b3b4ccdf113a6683630a0677cef928ed